+++ finish/openLibertyCafe/src/main/java/io/openliberty/guides/openlibertycafe/client/ServingWindowClient.java
@@ -10,7 +10,7 @@
  *     IBM Corporation - Initial implementation
  *******************************************************************************/
 // end::copyright[]
-package io.openliberty.guides.restaurantbff.client;
+package io.openliberty.guides.openlibertycafe.client;
 
 import org.eclipse.microprofile.openapi.annotations.tags.Tag;
 import org.eclipse.microprofile.rest.client.inject.RegisterRestClient;
