+++ finish/openLibertyCafe/src/main/liberty/config/server.xml
@@ -1,4 +1,4 @@
-<server description="Restaurant BFF">
+<server description="Open Liberty Cafe">
 
   <featureManager>
     <feature>beanValidation-2.0</feature>
@@ -15,5 +15,5 @@
 
   <httpEndpoint host="*" httpPort="${default.http.port}" httpsPort="${default.https.port}" id="defaultHttpEndpoint"/>
 
-  <webApplication location="restaurantBFF.war" contextRoot="/"/>
+  <webApplication location="openLibertyCafe.war" contextRoot="/"/>
 </server>
