+++ finish/scripts/startContainersBatScripts/openLibertyCafe.bat
@@ -1,5 +1,5 @@
 @ECHO OFF
-@ECHO starting the restaurantBFF container > restaurantBFF.tmp
+@ECHO starting the openLibertyCafe container > openLibertyCafe.tmp
 
 set ORDER_SERVICE_URL="http://order:9081"
 set SERVINGWINDOW_SERVICE_URL="http://servingwindow:9082"
@@ -9,8 +9,8 @@ docker run -d ^
   -e ServingWindowClient_mp_rest_url=%SERVINGWINDOW_SERVICE_URL% ^
   -p 9080:9080 ^
   --network=%NETWORK% ^
-  --name=restaurantbff ^
+  --name=openlibertycafe ^
   --rm ^
-  restaurantbff:1.0-SNAPSHOT
+  openlibertycafe:1.0-SNAPSHOT
 
-DEL restaurantBFF.tmp
+DEL openLibertyCafe.tmp
