+++ start/scripts/stopContainers.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
 
-docker stop kitchen bar order servingwindow restaurantbff kafka zookeeper
+docker stop kitchen bar order servingwindow openlibertycafe kafka zookeeper
 
 docker network rm reactive-app
